Added pagination size option and a higher-than-normal default.
As documented in https://github.com/chef-partners/vmware-vra-gem/issues/10,
there is a bug in the vRA API that causes duplicate entries to be returned
when the results are paginated.  This sets a high page size to avoid that,
but for users that have more than 200 items, this provides the ability
for the user to override the page size until VMware corrects the bug.